worked on optimal tap position for assignment 3
created class and added function to find optimal tap position based on criteria of minimizing line losses.

still need to code the part for the minimum voltage deviation criteria. I have some ideas how to do it, just gotta ask the professors something about this criteria, as its a bit vague in the description of assignment 3